+++ R/rxUiGet.R
@@ -284,6 +284,49 @@ rxUiGet.multipleEndpoint <- function(x, ...) {
 }
 attr(rxUiGet.multipleEndpoint, "desc") <- "table of multiple endpoint translations"
 
+#' This is a generic function for deparsing certain objects when
+#' printing out a rxode2 object.  Currently it is used for any meta-information
+#'
+#' @param object object to be deparsed
+#' @param var variable name to be assigned
+#' @return parsed R expression that can be used for printing and
+#'   `as.function()` calls.
+#' @export
+#' @author Matthew L. Fidler
+#' @examples
+#'
+#' mat <- matrix(c(1, 0.1, 0.1, 1), 2, 2, dimnames=list(c("a", "b"), c("a", "b")))
+#'
+#' rxUiDeparse(matrix(c(1, 0.1, 0.1, 1), 2, 2, dimnames=list(c("a", "b"), c("a", "b"))), "x")
+rxUiDeparse <- function(object, var) {
+ UseMethod("rxUiDeparse")
+}
+
+#' @rdname rxUiDeparse
+#' @export
+rxUiDeparse.lotriFix <- function(object, var) {
+  .val <- lotri::lotriAsExpression(object)
+  bquote(.(str2lang(var)) <- .(.val))
+}
+
+#' @rdname rxUiDeparse
+#' @export
+rxUiDeparse.default <- function(object, var) {
+  # This is a default method for deparsing objects
+  if (checkmate::testMatrix(object, any.missing=FALSE,
+                            row.names="strict", col.names="strict")) {
+    .dn <- dimnames(object)
+    if (identical(.dn[[1]], .dn[[2]]) && isSymmetric(object)) {
+      return(rxUiDeparse.lotriFix(object, var))
+    }
+  }
+  .ret <- try(str2lang(paste0(var, "<-", deparse1(object))))
+  if (inherits(.ret, "try-error")) {
+    .ret <- str2lang("NULL")
+  }
+  .ret
+}
+
 #' @rdname rxUiGet
 #' @export
 rxUiGet.funPrint <- function(x, ...) {
@@ -295,20 +338,7 @@ rxUiGet.funPrint <- function(x, ...) {
   for (.i in seq_along(.ls)) {
     .var <- .ls[.i]
     .val <- .x$meta[[.ls[.i]]]
-    .isLotri <- FALSE
-    if (checkmate::testMatrix(.val, any.missing=FALSE, row.names="strict", col.names="strict")) {
-      .dn <- dimnames(.val)
-      if (identical(.dn[[1]], .dn[[2]]) && isSymmetric(.val)) {
-        class(.val) <- c("lotriFix", class(.val))
-        .val <- as.expression(.val)
-        .val <- bquote(.(str2lang(.var)) <- .(.val))
-        .ret[[.i + 1]] <- .val
-        .isLotri <- TRUE
-      }
-    }
-    if (!.isLotri) {
-      .ret[[.i + 1]] <- eval(parse(text=paste("quote(", .var, "<-", deparse1(.val), ")")))
-    }
+    .ret[[.i + 1]] <- rxUiDeparse(.val, .var)
   }
   .theta <- x$theta
   .omega <- x$omega